65 afinar UI de pivot table section

+++ python-app/lib/pivot_table/plot_pivot_table.py
@@ -3,7 +3,7 @@ from models.pivot_table.pivot_table_level import PivotTableLevel, level_to_spani
 from models.pivot_table.aggregate_function_type import AggregateFunctionType, aggregate_function_to_spanish
 from models.pivot_table.filter_function_type import FilterFunctionType, filter_function_to_spanish
 from models.pivot_table.data_filter.charting_mode import ChartingMode
-from models.report import Report
+from models.report.self import Report
 
 from lib.pivot_table.plot_data import plot_data
 from lib.directory_definitions import preview_image_of_slide
@@ -28,24 +28,20 @@ def plot_from_components(
         filepath=filepath
         )
     
-    if former_preview is not None:
+    if former_preview is not None and os.path.exists(former_preview):
         os.remove(former_preview)
 
-def plot_from_entities(report: Report, pivot_table: PivotTable) -> str:
+def plot_from_entities(filepath: str, pivot_table: PivotTable) -> str:
     outer_filter = next((_filter for _filter in pivot_table.filters if _filter.charting_mode == ChartingMode.SUPER_CHART), None)
     if outer_filter is None:
         outer_filter = next((_filter for _filter in pivot_table.filters if _filter.charting_mode == ChartingMode.CHART), None)
 
-    filepath = preview_image_of_slide(root_directory=report.root_directory, slide_id=pivot_table.identifier)
-
     plot_from_components(
         data=pivot_table.data, 
-        title=pivot_table.name, 
+        title=pivot_table.title, 
         outer_chart=outer_filter.level, 
         _filter=pivot_table.filter_function, 
         aggregate=pivot_table.aggregate_function, 
         filepath=filepath, 
         former_preview=pivot_table.preview
         )
-    
-    return filepath
